♻️ REFACTOR: service 계층 엔티티 단위로 폴더 정리
