+++ src/main/java/cloudcomputinginha/demo/service/interview/InterviewQueryService.java
@@ -1,9 +1,11 @@
-package cloudcomputinginha.demo.service;
+package cloudcomputinginha.demo.service.interview;
 
 import cloudcomputinginha.demo.web.dto.InterviewResponseDTO;
+
 import java.util.List;
 
 public interface InterviewQueryService {
     List<InterviewResponseDTO.InterviewGroupCardDTO> getGroupInterviewCards();
+
     InterviewResponseDTO.GroupInterviewDetailDTO getGroupInterviewDetail(Long interviewId);
 }
