+++ src/main/java/cloudcomputinginha/demo/service/memberInterview/MemberInterviewCommandService.java
@@ -1,4 +1,4 @@
-package cloudcomputinginha.demo.service;
+package cloudcomputinginha.demo.service.memberInterview;
 
 import cloudcomputinginha.demo.domain.MemberInterview;
 import cloudcomputinginha.demo.domain.enums.InterviewStatus;
