+++ src/main/java/cloudcomputinginha/demo/service/memberInterview/MemberInterviewSocketService.java
@@ -1,10 +1,11 @@
-package cloudcomputinginha.demo.service;
+package cloudcomputinginha.demo.service.memberInterview;
 
 import cloudcomputinginha.demo.web.dto.MemberInterviewSocketMessageDTO.WaitingRoomActionDTO;
-import java.util.List;
 
 public interface MemberInterviewSocketService {
     void enterWaitingRoom(WaitingRoomActionDTO waitingRoomActionDTO, String sessionId);
+
     void leaveWaitingRoom(Long interviewId, Long memberId, String sessionId);
+
     void enterInterview(Long interviewId);
 }
