+++ src/main/java/cloudcomputinginha/demo/service/resume/ResumeQueryServiceImpl.java
@@ -1,4 +1,4 @@
-package cloudcomputinginha.demo.service;
+package cloudcomputinginha.demo.service.resume;
 
 import cloudcomputinginha.demo.domain.Resume;
 import cloudcomputinginha.demo.repository.ResumeRepository;
