+++ src/main/java/cloudcomputinginha/demo/service/resume/ResumeS3Service.java
@@ -1,4 +1,4 @@
-package cloudcomputinginha.demo.service;
+package cloudcomputinginha.demo.service.resume;
 
 import cloudcomputinginha.demo.apiPayload.code.handler.ResumeHandler;
 import cloudcomputinginha.demo.apiPayload.code.status.ErrorStatus;
@@ -15,7 +15,7 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class S3PresignedService {
+public class ResumeS3Service {
     private final S3Presigner s3Presigner;
 
     @Value("${cloud.aws.s3.bucket}")
