+++ src/main/java/cloudcomputinginha/demo/web/controller/CoverletterRestController.java
@@ -4,9 +4,9 @@ import cloudcomputinginha.demo.apiPayload.ApiResponse;
 import cloudcomputinginha.demo.converter.CoverletterConverter;
 import cloudcomputinginha.demo.domain.Coverletter;
 import cloudcomputinginha.demo.domain.Qna;
-import cloudcomputinginha.demo.service.CoverletterCommandService;
-import cloudcomputinginha.demo.service.CoverletterQueryService;
-import cloudcomputinginha.demo.service.QnaQueryService;
+import cloudcomputinginha.demo.service.coverletter.CoverletterCommandService;
+import cloudcomputinginha.demo.service.coverletter.CoverletterQueryService;
+import cloudcomputinginha.demo.service.qna.QnaQueryService;
 import cloudcomputinginha.demo.validation.annotation.ExistCoverletter;
 import cloudcomputinginha.demo.validation.annotation.ExistMember;
 import cloudcomputinginha.demo.web.dto.CoverletterRequestDTO;
