+++ src/main/java/cloudcomputinginha/demo/web/controller/InterviewRestController.java
@@ -4,9 +4,9 @@ package cloudcomputinginha.demo.web.controller;
 import cloudcomputinginha.demo.apiPayload.ApiResponse;
 import cloudcomputinginha.demo.converter.InterviewConverter;
 import cloudcomputinginha.demo.domain.Interview;
-import cloudcomputinginha.demo.service.InterviewCommandService;
-import cloudcomputinginha.demo.service.InterviewOptionCommandService;
-import cloudcomputinginha.demo.service.InterviewQueryService;
+import cloudcomputinginha.demo.service.interview.InterviewCommandService;
+import cloudcomputinginha.demo.service.interview.InterviewQueryService;
+import cloudcomputinginha.demo.service.interviewOption.InterviewOptionCommandService;
 import cloudcomputinginha.demo.validation.annotation.ExistInterview;
 import cloudcomputinginha.demo.web.dto.InterviewOptionRequestDTO;
 import cloudcomputinginha.demo.web.dto.InterviewOptionResponseDTO;
